requirements.txt
# ======================================================
# Python version recommendation (NOT a pip package)
# ======================================================
# Use Python 3.10 or 3.11

# ======================================================
# Core numerical stack
# ======================================================
numpy>=1.23
scipy>=1.10

# ======================================================
# Image processing
# ======================================================
scikit-image>=0.21
tifffile>=2023.7.0

# ======================================================
# Visualization
# ======================================================
matplotlib>=3.7
napari>=0.4.18
vispy>=0.14        # required backend for napari

# ======================================================
# GUI support (napari + manual ROI)
# ======================================================
PyQt5>=5.15

# ======================================================
# File I/O
# ======================================================
h5py>=3.9
PyYAML>=6.0

# ======================================================
# Video generation
# ======================================================
imageio>=2.31
imageio-ffmpeg>=0.4.8

# ======================================================
# Optional (performance, future scaling)
# ======================================================
numba>=0.57
